+++ humble_ws/src/isaac_tutorials/scripts/ros2_ackermann_publisher.py
@@ -8,11 +8,15 @@
 # distribution of this software and related documentation without an express
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 
+import signal
+import sys
+
+import numpy as np
+
 import rclpy
 from rclpy.node import Node
 
 from ackermann_msgs.msg import AckermannDriveStamped
-import numpy as np
 
 class MinimalPublisher(Node):
 
@@ -25,19 +29,27 @@ class MinimalPublisher(Node):
 
     def timer_callback(self):
         msg = AckermannDriveStamped()
-        
-        # Only command forklift using acceleration and steering angle
+
+        # Only command ackermann drive robot using speed and steering angle
         degrees1 = np.arange(0, 60)
         degrees2 = np.arange(-60, 0)
         degrees = np.concatenate((degrees1, degrees1[::-1], degrees2[::-1], degrees2))
-        msg.header.frame_id = "forklift"
+        msg.header.frame_id = "ackermann"
         msg.header.stamp = self.get_clock().now().to_msg()
         msg.drive.steering_angle =  0.0174533 * (degrees[self.i % degrees.size])
-        
-        msg.drive.acceleration = float(degrees[self.i % degrees.size])
-        
+
+        msg.drive.speed = float(degrees[self.i % degrees.size])
+        msg.drive.acceleration = 1.0
         self.publisher_.publish(msg)
         self.i += 1
+
+    def stop_robot(self):
+        msg = AckermannDriveStamped()
+
+        # Publish zero-velocity message
+        msg.drive.steering_angle = 0.0
+        msg.drive.speed = 0.0
+        self.publisher_.publish(msg)  
 
 
 def main(args=None):
@@ -46,10 +58,19 @@ def main(args=None):
 
     minimal_publisher = MinimalPublisher()
 
-    rclpy.spin(minimal_publisher)
+    def signal_handler(sig, frame):
+        minimal_publisher.stop_robot()
+        minimal_publisher.destroy_node()
+        rclpy.shutdown()
+        sys.exit(0)
 
-    minimal_publisher.destroy_node()
-    rclpy.shutdown()
+    # Register the SIGINT handler for Ctrl+C
+    signal.signal(signal.SIGINT, signal_handler)
+
+    try:
+        rclpy.spin(minimal_publisher)
+    except KeyboardInterrupt:
+        pass
 
 
 if __name__ == '__main__':