+++ humble_ws/src/isaacsim/launch/run_isaacsim.launch.py
@@ -16,7 +16,7 @@ from launch_ros.substitutions import FindPackageShare
 
 # Declare all launch arguments corresponding to the bash script options
 launch_args = [
-    DeclareLaunchArgument('version', default_value='4.2.0', description='Specify the version of Isaac Sim to use. Isaac Sim will be run from default install root folder for the specified version. Leave empty to use latest version of Isaac Sim.'),
+    DeclareLaunchArgument('version', default_value='4.5.0', description='Specify the version of Isaac Sim to use. Isaac Sim will be run from default install root folder for the specified version. Leave empty to use latest version of Isaac Sim.'),
     
     DeclareLaunchArgument('install_path', default_value='', description='If Isaac Sim is insalled in a non-default location, provide a specific path to Isaac Sim installation root folder. (If defined, "version" parameter will be ignored)'),
     
@@ -35,6 +35,8 @@ launch_args = [
     DeclareLaunchArgument('ros_installation_path', default_value='', description='If ROS is installed in a non-default location (as in not under /opt/ros/), provide the path to your main setup.bash file for your ROS install. (/path/to/custom/ros/install/setup.bash)'),
 
     DeclareLaunchArgument('headless', default_value='', description='Set to "native" or "webrtc" to run Isaac Sim with different headless modes, if left blank, Isaac Sim will run in the regular GUI workflow. This parameter can be overridden by "standalone" parameter.'),
+
+    DeclareLaunchArgument('custom_args', default_value='', description='Add any custom Isaac Sim args that you want to forward to isaac-sim.sh during run time.'),
 
 ]
 
@@ -79,7 +81,8 @@ def launch_setup(context):
             'play_sim_on_start': LaunchConfiguration('play_sim_on_start'),
             'ros_distro': LaunchConfiguration('ros_distro'),
             'ros_installation_path': LaunchConfiguration('ros_installation_path'),
-            'headless': LaunchConfiguration('headless')
+            'headless': LaunchConfiguration('headless'),
+            'custom_args': LaunchConfiguration('custom_args')
         }]
     )
     return [isaacsim_node]